+++ src/components/main/header/SearchHeader.module.css
@@ -1,5 +1,16 @@
 header {
     display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    width: 100%;
+    height: 4.375rem;
+
+    border-bottom: 1px solid #bebebe;
+
+    padding: 1.25rem 0;
+
+    margin-bottom: 2.5rem;
 }
 
 .logo {
