+++ src/components/main/list/VideoList.jsx
@@ -1,10 +1,13 @@
 import React from 'react';
 import VideoCard from './item/VideoCard';
+import styles from './VideoList.module.css';
 
 export default function VideoList() {
     return (
-        <div>
-            <VideoCard />
+        <div className={styles.container}>
+            {new Array(20).fill(1).map((item, idx) => (
+                <VideoCard key={idx} />
+            ))}
         </div>
     );
 }
