+++ src/components/main/list/VideoList.module.css
@@ -1,0 +1,8 @@
+.container {
+    display: grid;
+    grid-template-columns: repeat(5, 1fr);
+
+    width: 100%;
+
+    gap: 0.5rem;
+}
