+++ src/components/main/list/item/VideoCard.module.css
@@ -1,0 +1,34 @@
+.card {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+
+    min-width: 90%;
+
+    color: var(--color-light-gray);
+    font-size: 0.75rem;
+}
+
+.thumbnail {
+    width: 100%;
+    height: 10rem;
+
+    background-color: lightblue;
+    margin-bottom: 0.5rem;
+}
+
+.title {
+    color: var(--color-text);
+    font-weight: 700;
+    font-size: 1rem;
+
+    margin-bottom: 0.625rem;
+}
+
+.name {
+    margin-bottom: 0.2rem;
+}
+
+.time {
+    margin-bottom: 0.625rem;
+}
