+++ src/pages/Pages.module.css
@@ -1,0 +1,6 @@
+.page {
+    display: flex;
+    flex-direction: column;
+
+    width: 90%;
+}
